reactor: 상수에서 조회하는 걸로 변경

+++ src/main/java/lotto/domain/lotto/Lottos.java
@@ -79,7 +79,7 @@ public class Lottos {
 
     private long prizeTotal(List<Winning> winnings) {
         return winnings.stream()
-                .mapToInt(Winning::prize)
+                .mapToLong(Winning::prize)
                 .sum();
     }
